+++ packages/network/package.json
@@ -28,6 +28,7 @@
     "test-unit": "jest --forceExit test/unit",
     "coverage": "jest --forceExit --coverage",
     "test-integration": "jest --forceExit test/integration",
+    "test-flakey": "jest --forceExit test/flakey",
     "test-types": "tsc --noEmit --project ./tsconfig.json ",
     "eslint": "eslint .",
     "network": "node $NODE_DEBUG_OPTION bin/network.js",
